# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Documentation for API endpoints",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/checklists": {
            "get": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Ambil semua checklist user (admin bisa semua)",
                "operationId": "348d0ba4f956147227c4e0f6574b8573",
                "responses": {
                    "200": {
                        "description": "List checklist dikembalikan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Buat checklist baru",
                "operationId": "6cd3aa07d0b04a7619eeddde30be97bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "due_time",
                                    "repeat_interval"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Olahraga"
                                    },
                                    "due_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-07-30T10:00:00"
                                    },
                                    "repeat_interval": {
                                        "type": "string",
                                        "enum": [
                                            "daily",
                                            "3_days",
                                            "weekly",
                                            "monthly",
                                            "yearly"
                                        ],
                                        "example": "weekly"
                                    },
                                    "repeat_days": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "monday",
                                                "tuesday",
                                                "wednesday",
                                                "thursday",
                                                "friday",
                                                "saturday",
                                                "sunday"
                                            ]
                                        },
                                        "example": [
                                            "sunday",
                                            "friday"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checklist dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checklists/{id}": {
            "get": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Lihat detail checklist berdasarkan ID",
                "operationId": "0b6145970b9b33ac87c035bafa697e4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID checklist",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail checklist dikembalikan"
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Checklist tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Update checklist",
                "operationId": "ceefd03f627edf699b4000d511899551",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID checklist",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Olahraga Pagi"
                                    },
                                    "due_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-08-01T07:00:00"
                                    },
                                    "repeat_interval": {
                                        "type": "string",
                                        "enum": [
                                            "daily",
                                            "3_days",
                                            "weekly",
                                            "monthly",
                                            "yearly"
                                        ],
                                        "example": "weekly"
                                    },
                                    "repeat_days": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "monday",
                                                "tuesday",
                                                "wednesday",
                                                "thursday",
                                                "friday",
                                                "saturday",
                                                "sunday"
                                            ]
                                        },
                                        "example": [
                                            "tuesday",
                                            "thursday"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checklist diperbarui"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Checklist tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Hapus checklist (soft delete)",
                "operationId": "1e20f3721b541582dee802bd3db4f1bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID checklist",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist dihapus"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Checklist tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checklists/{id}/restore": {
            "post": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Mengembalikan checklist yang telah dihapus (soft delete)",
                "operationId": "7ac167bc2cf58e348481c64bc5285ad8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID checklist",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist berhasil dikembalikan"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Checklist tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checklists/{id}/complete": {
            "post": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Tandai checklist sebagai selesai (beserta repeat days)",
                "description": "Menandai checklist utama dan semua repeat_days sebagai selesai.",
                "operationId": "88e53d2d836adaeeb66f347af26c0942",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID checklist yang ingin ditandai selesai",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist ditandai selesai"
                    },
                    "403": {
                        "description": "Tidak diizinkan"
                    },
                    "404": {
                        "description": "Checklist tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checklists/{id}/uncomplete": {
            "post": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Tandai checklist sebagai belum selesai (beserta repeat days)",
                "description": "Menandai checklist utama dan semua repeat_days sebagai belum selesai.",
                "operationId": "e4ab46c312c945e8ed42f04ec24d5676",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID checklist yang ingin ditandai belum selesai",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist ditandai belum selesai"
                    },
                    "403": {
                        "description": "Tidak diizinkan"
                    },
                    "404": {
                        "description": "Checklist tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/checklists/{id}/repeat-days/{day}/complete": {
            "post": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Tandai checklist mingguan di hari tertentu sebagai selesai",
                "operationId": "c3e9a2eca0e875f7a49a95daa4d28529",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID checklist",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "day",
                        "in": "path",
                        "description": "Hari dalam minggu (e.g. monday, tuesday, dst)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "monday",
                                "tuesday",
                                "wednesday",
                                "thursday",
                                "friday",
                                "saturday",
                                "sunday"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hari checklist ditandai selesai"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT Anda",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}